pharmacy_backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-cors-headers
python-decouple
pillow